feat: Add make request action & code refactor
