+++ src/actions/get_object.py
@@ -1,0 +1,47 @@
+# Copyright (c) 2026 Splunk Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from soar_sdk.abstract import SOARClient
+from soar_sdk.action_results import ActionOutput, OutputField
+from soar_sdk.params import Param, Params
+
+from ..asset import Asset
+from ..salesforce_client import SalesforceClient
+
+
+class GetObjectParams(Params):
+    sobject: str = Param(
+        description="Name of object",
+        primary=True,
+        default="Case",
+        cef_types=["salesforce object name"],
+    )
+    id: str = Param(
+        description="Salesforce Object ID",
+        primary=True,
+        cef_types=["salesforce object id"],
+    )
+
+
+class GetObjectOutput(ActionOutput):
+    id: str = OutputField(
+        cef_types=["salesforce object id"], example_values=["5001I000002SfMMQA0"]
+    )
+
+
+def get_object(
+    params: GetObjectParams, soar: SOARClient, asset: Asset
+) -> GetObjectOutput:
+    record = SalesforceClient(asset).get(params.sobject, params.id)
+    soar.set_message(f"Successfully retrieved {params.sobject}")
+    return GetObjectOutput(id=record.get("Id", params.id))
